// src/app/api/models/customer.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Customer {

  /**
   * Multiline.
   */
  address?: string | null;
  created_at?: string;
  email?: string | null;
  id?: string;
  name?: string;

  /**
   * Private notes, not shown on documents.
   */
  notes?: string | null;
  phone?: string | null;
  updated_at?: string;
}